refactor(plugin-structure): 重组插件项目结构，将主动能力相关模块移动到 proactive 目录，将 Web UI 相关模块移动到 web 目录，并更新导入路径
This reorganization separates proactive modules (memory API gateway, topic engine, user profiling, temporal memory, memory events) and web UI modules under dedicated packages, updating imports and initialization flows accordingly. It improves maintainability and plugin development ergonomics.

- Introduce proactive and web packages and adjust imports across core components
- Add package-level __init__.py for exported modules and clearer imports
- Update documentation references to reflect new structure

BREAKING CHANGE: Import paths for MemoryWebServer and proactive modules changed; update your plugin code to import from the new package locations (e.g., from web.web_server import MemoryWebServer, from proactive.memory_api_gateway import MemoryAPIGateway).
