+++ proactive/__init__.py
@@ -1,0 +1,1 @@
+"""主动能力相关模块：话题引擎、用户画像、时间维度记忆和统一API网关等。"""
