+++ web/__init__.py
@@ -1,0 +1,1 @@
+"""Web 相关模块：内置管理 Web 服务及静态资源。"""
