+++ web/web_server.py
@@ -14,7 +14,7 @@ except Exception:  # pragma: no cover
     import logging
     logger = logging.getLogger(__name__)
 
-from .resource_management import resource_manager
+from ..resource_management import resource_manager
 from .web_assets import DEFAULT_INDEX_HTML, DEFAULT_STYLE_CSS, DEFAULT_APP_JS
